+++ reporting/rep602.php
@@ -133,12 +133,12 @@ function print_bank_transactions_reconcile(): void
 				if ($myrow['amount'] > 0.0)
 				{
 					$rep->AmountCol(5, 6, abs($myrow['amount']), $dec);
-					$total_debit += abs($myrow['amount']);
+					$total_debit += (float)abs($myrow['amount']);
 				}
 				else
 				{
 					$rep->AmountCol(6, 7, abs($myrow['amount']), $dec);
-					$total_credit += abs($myrow['amount']);
+					$total_credit += (float)abs($myrow['amount']);
 				}
 				$rep->AmountCol(7, 8, $total, $dec);
 				if ((bool)$myrow["reconciled"] && $myrow["reconciled"] != '0000-00-00')
